Update UI and personality
