+++ src/main/java/chick/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package chick;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -32,7 +32,6 @@ public class DialogBox extends HBox {
         } catch (IOException e) {
             e.printStackTrace();
         }
-
         dialog.setText(text);
         displayPicture.setImage(img);
     }
