+++ src/main/java/chick/Launcher.java
@@ -1,4 +1,4 @@
-package duke;
+package chick;
 
 import javafx.application.Application;
