+++ src/main/java/chick/Main.java
@@ -1,4 +1,4 @@
-package duke;
+package chick;
 
 import java.io.IOException;
