+++ src/main/java/chick/task/DeadlineTask.java
@@ -1,30 +1,30 @@
-package duke.task;
+package chick.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import duke.ChickException;
+import chick.ChickException;
 
 /**
- * Class representing an EventTask.
+ * Class representing a DeadlineTask.
  */
-public class EventTask extends Task {
-    protected LocalDate at;
+public class DeadlineTask extends Task {
+    protected LocalDate by;
 
     /**
-     * Class constructor for EventTask.
+     * Class constructor for DeadlineTask.
      *
-     * @param description Command string being used to create EventTask.
+     * @param description Command string being used to create DeadlineTask.
      * @throws DateTimeParseException If datetime given cannot be parsed.
      * @throws ChickException If command is invalid.
      */
-    public EventTask(String description) throws DateTimeParseException, ChickException {
+    public DeadlineTask(String description) throws DateTimeParseException, ChickException {
         super();
         this.commandString = description;
-        int descriptionStartIndex = "event ".length();
+        int descriptionStartIndex = "deadline ".length();
         description = description.substring(descriptionStartIndex);
-        String[] split = description.split(" /at ");
+        String[] split = description.split(" /by ");
         int correctSplitCount = 2;
         if (split.length < correctSplitCount) {
             throw new ChickException("Deadline time (indicated by /by separator) is missing.");
@@ -32,12 +32,11 @@ public class EventTask extends Task {
             throw new ChickException("Multiple usage of /by separator is not allowed.");
         }
         this.description = split[0];
-        this.at = LocalDate.parse(split[1]);
+        this.by = LocalDate.parse(split[1]);
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " + at.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
+        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
     }
-
 }
