+++ src/main/java/chick/task/EventTask.java
@@ -1,30 +1,30 @@
-package duke.task;
+package chick.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import duke.ChickException;
+import chick.ChickException;
 
 /**
- * Class representing a DeadlineTask.
+ * Class representing an EventTask.
  */
-public class DeadlineTask extends Task {
-    protected LocalDate by;
+public class EventTask extends Task {
+    protected LocalDate at;
 
     /**
-     * Class constructor for DeadlineTask.
+     * Class constructor for EventTask.
      *
-     * @param description Command string being used to create DeadlineTask.
+     * @param description Command string being used to create EventTask.
      * @throws DateTimeParseException If datetime given cannot be parsed.
      * @throws ChickException If command is invalid.
      */
-    public DeadlineTask(String description) throws DateTimeParseException, ChickException {
+    public EventTask(String description) throws DateTimeParseException, ChickException {
         super();
         this.commandString = description;
-        int descriptionStartIndex = "deadline ".length();
+        int descriptionStartIndex = "event ".length();
         description = description.substring(descriptionStartIndex);
-        String[] split = description.split(" /by ");
+        String[] split = description.split(" /at ");
         int correctSplitCount = 2;
         if (split.length < correctSplitCount) {
             throw new ChickException("Deadline time (indicated by /by separator) is missing.");
@@ -32,11 +32,12 @@ public class DeadlineTask extends Task {
             throw new ChickException("Multiple usage of /by separator is not allowed.");
         }
         this.description = split[0];
-        this.by = LocalDate.parse(split[1]);
+        this.at = LocalDate.parse(split[1]);
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
+        return "[E]" + super.toString() + " (at: " + at.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
     }
+
 }
