+++ src/main/resources/view/DialogBox.fxml
@@ -7,8 +7,8 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="600.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" text="Label" wrapText="true"/>
-        <ImageView fx:id="displayPicture" fitHeight="200.0" fitWidth="155.0" pickOnBounds="true" preserveRatio="true" />
+        <Label fx:id="dialog" text="Label" wrapText="true" minHeight="-Infinity"/>
+        <ImageView fx:id="displayPicture" fitHeight="99" fitWidth="99" pickOnBounds="true" preserveRatio="true" />
     </children>
     <padding>
         <Insets bottom="15.0" left="5.0" right="5.0" top="15.0" />
