+++ src/test/java/chick/TaskListTest.java
@@ -1,4 +1,4 @@
-package duke;
+package chick;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
